modification of the namespace `CwAPI3D_Net::Utils` to `CwAPI3D::Net::Bridge::Utils` to reflect the new project structure, the change of `elementId` type from `int` to `uint64_t` to accommodate larger values, and the addition of a new method `createRectangularBeamPoints` that creates a rectangular beam with given parameters.

+++ CwAPI3D.Net.Bridge/src/typedefs.h
@@ -1,3 +1,3 @@
 #pragma once
 
-using elementId = int;
+using elementId = uint64_t;
